requirements.txt
async-timeout==4.0.2
attrs==22.2.0
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.0.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
exceptiongroup==1.1.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.0
pluggy==1.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==7.2.2
python-apt==2.4.0+ubuntu1
PyYAML==5.4.1
redis==4.5.1
requests==2.28.2
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
tomli==2.0.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
wadllib==1.3.6
zipp==1.0.0
